% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HueLightFunctions.R
\name{alertLights}
\alias{alertLights}
\title{Make the lights blink as an allerting effect}
\usage{
alertLights(ip, userName, light)
}
\arguments{
\item{ip}{IP address generated by getIP()}

\item{userName}{A username key}

\item{light}{An available light number}
}
\value{
Have the lights alert
}
\description{
Make the lights blink as an allerting effect
}
\examples{
alertLights("0.0.0.0", "XXXX", 1)
}
