% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HueLightFunctions.R
\name{changeBrightness}
\alias{changeBrightness}
\title{Change Brightness of a light}
\usage{
changeBrightness(ip, userName, light, brightness)
}
\arguments{
\item{ip}{IP address generated by getIP()}

\item{userName}{A username key}

\item{light}{An available light number}

\item{brightness}{Brightness between (1 - 254)}
}
\value{
Changes the brightness of the light
}
\description{
Change Brightness of a light
}
\examples{
changeBrightness("0.0.0.0","Xo12x",1,255)
}
