% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HueLightFunctions.R
\name{changeEffect}
\alias{changeEffect}
\title{Change light effect}
\usage{
changeEffect(ip, userName, light, effect)
}
\arguments{
\item{ip}{IP address generated by getIP()}

\item{userName}{A username key}

\item{light}{An available light number}

\item{effect}{}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Change light effect
}
\examples{
changeEffect("0.0.0.0","XXXX",1,"Carsol")
}
