% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HueLightFunctions.R
\name{changeHue}
\alias{changeHue}
\title{Change hue of a light}
\usage{
changeHue(ip, userName, light, hue)
}
\arguments{
\item{ip}{IP address generated by getIP()}

\item{userName}{A username key}

\item{light}{An available light number}

\item{hue}{Hue between (0 - 65535)}
}
\value{
Changes the hue of the light
}
\description{
Change hue of a light
}
\examples{
changeHue(ip,userName,light,"254")
}
