% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HueLightFunctions.R
\name{getLights}
\alias{getLights}
\title{Creates a list of available lights}
\usage{
getLights(ip, userName)
}
\arguments{
\item{ip}{IP address generated by getIP()}

\item{userName}{A username key}
}
\value{
Returns a list of lights to be changed
}
\description{
Creates a list of available lights
}
